rename errorReducer, add redirect action, fix search rendering image issue
