+++ src/components/Roomie/RoomieBox.js
@@ -8,15 +8,12 @@ import img3 from '../../styles/img/person-3.jpg';
 export class RoomieBox extends React.Component {
     constructor(props) {
         super(props);
-        this.state = {
-            profileImage: "/api/image/" + this.props.roomie.profileImage
-        }
     }
 
     render() {
         return (
             <a href="#" className="roomie-box">
-                <img src={this.state.profileImage} alt="person" className="roomie-box__img" />
+                <img src={`/api/image/${this.props.roomie.profileImage}`} alt="person" className="roomie-box__img" />
                 <p className="roomie-box__name">{this.props.roomie.name}</p>
                 <p className="roomie-box__address">{this.props.roomie.region}</p>
                 <button className="button button--primary">More</button>
