+++ src/components/Roomie/RoomieRegister.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 import * as actions from '../../redux/actions/roomie';
 import FormData from 'form-data';
+import { Redirect } from 'react-router-dom';
 
 class RoomieRegister extends React.Component {
     constructor() {
@@ -41,6 +42,9 @@ class RoomieRegister extends React.Component {
     render() {
         const { phoneNumber, region, minBudget, maxBudget } = this.state;
 
+        if(this.props.status.redirect == "home") {
+           return <Redirect to="/home" />
+        }
         return (
             <div className="register">
                 <form onSubmit={this.handleSubmit} className="register-form">
@@ -94,7 +98,8 @@ class RoomieRegister extends React.Component {
 
 function mapStateToProps(state) {
     return {
-        auth: state.authReducer
+        auth: state.authReducer,
+        status: state.statusReducer
     }
 }
