+++ src/components/Sidebar/SidebarProfie.js
@@ -4,7 +4,7 @@ export class SidebarProfile extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
-            profileImage: "/api/profileimage/" + this.props.auth.user.roomie.profileImage
+            profileImage: "/api/image/" + this.props.auth.user.roomie.profileImage
         }
     }
     // replace this value with state
