+++ src/redux/actionTypes.js
@@ -1,6 +1,7 @@
 export const SET_CURRENT_USER = "SET_CURRENT USER";
 
 export const SET_ERROR = "SET_ERROR";
+export const SET_REDIRECT = "SET_REDIRECT";
 
 export const GET_ROOMIES = "GET_ROOMIES";
