--- src/redux/actions/error.js
@@ -1,6 +1,0 @@
-import { SET_ERROR } from '../actionTypes';
-
-export const setError = error => ({
-    type: SET_ERROR,
-    error
-});
